Refactor signal visualization
